Auto-correct docker source URI
Skopeo does not accept "docker:///" as valid URI format. [1]
However, driver URIs within the Libvirt project and virtualization
tools such as libvirt-sandbox use three slashes after URI scheme. [2]

Make virt-bootstrap auto-correct the source URI for consistency with
other tools.

Following the syntax specifications in RFC 1808, "urlparse" recognizes
a netloc only if it is properly introduced by ‘//’. [3] Otherwise the
input is presumed to be a relative URL and thus to start with a path
component. [4]

- "self.image" is the path component
- "self.registry" is the netloc component

[1] https://github.com/projectatomic/skopeo/blob/master/docs/skopeo.1.md#skopeo-copy-1
[2] http://libvirt.org/uri.html
[3] https://docs.python.org/2/library/urlparse.html
[4] https://hg.python.org/cpython/file/2.7/Lib/urlparse.py#l187