Python 3/2 compatibility: Convert Byte-str to Str
Encoded Unicode in Python 3 is represented as binary data. The
difference with Python2 is that any attempt to mix text and data in
Python 3.0 raises TypeError, whereas if you were to mix Unicode and
8-bit strings in Python 2.x, it would work if the 8-bit string happened
to contain only 7-bit (ASCII) bytes, but you would get
UnicodeDecodeError if it contained non-ASCII values.

Reference:
https://docs.python.org/release/3.0.1/whatsnew/3.0.html#text-vs-data-instead-of-unicode-vs-8-bit

Example:

Python 2:

    >>> b'foo.bar'.split('.')
    ['foo', 'bar']

Python 3:

    >>> b'foo.bar'.split('.')
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    TypeError: a bytes-like object is required, not 'str'

    >>> b'foo.bar'.split(b'.')
    [b'foo', b'bar']