Drop need of root privileges to set root password
These changes aim to avoid the requirement for root privileges when
setting the password of root user on root file system.

The "-R, --root" flag of chpasswd is using chroot to apply changes in
root file system and this requires root privileges. [1]

Instead compute hash of the root password using passlib [2] and insert
the value in the /etc/shadow file in the rootfs.

[1] https://en.wikipedia.org/wiki/Chroot#Limitations
[2] http://passlib.readthedocs.io/en/stable/lib/passlib.hosts.html