Use requirements.txt for dependencies
Later on when we will introduce libguestfs dependency, we would need
to add a dependency_links to setup.py. This is rather hacky and
https://caremad.io/posts/2013/07/setup-vs-requirement/ recommends
using requirements.txt.

Thus, to install the package when all dependencies are already
installed, the following command can be used:

    python setup.py install

Otherwise, the recommended way is to use pip as follows:

    pip install -r recommends.txt -e .

This will also download the guestfs python binding, build and install
it. However this would be skipped if installed from a linux distro
package if this one provides the egg-info.