Improve untar command
Add "--overwrite" to enforce the overwrite of existing files.

Add the flag "--absolute-names" to disable the strip of leading '/'s
This is used to get around the error "Cannot open:Permission denied"
which occurs when the qemu driver is used by virt-sandbox. It is used
for unprivileged users to create isolated environment in which tar is
executed to extract the content from container image layers.

In particular this error occurs when the tar archive contains symbolic
link which has target path starting with '/'.

Steps to reproduce:
    $ mkdir /tmp/foo
    $ cd /tmp/foo
    $ touch file
    $ ln -s /tmp/foo/file link
    $ tar -cf archive.tar link
    $ mkdir /tmp/foo/dest
    $ virt-sandbox -c qemu:///session \
                   -m host-bind:/mnt=/tmp/foo/dest \
                   -- /bin/tar xf /tmp/foo/archive.tar -C /mnt

Error message:
    tar: link: Cannot open: Permission denied
    tar: Exiting with failure status due to previous errors