Create qcow2 images with guestfs-python
Use the python bindings of libguestfs to create qcow2 image with
backing chains to mimic the layers of container image.

This commit also changes the behavior of FileSource when 'qcow2'
output format is used. Now the string layer-0.qcow2 will be used
as name of the output file.

This change is applied in the test suite as an update to the function
get_image_path().