docker: only one layer drive per guestfs appliance
The current code was trying to save time by adding all drives to the
guestfs handle and launch the instance from it. However, this doesn't
fly since backing chain images would be used more than once.

To bad for performance, but sticking to one layer per guestfs launch
is safer and gets the tests to the pass.