docker-source: Get list of layers without `--raw`
When `skopeo inspect --raw docker://feodra` is used the returned
manifest content contains a list with manifests for specific
platforms [1] rather than a list with layers.

By using `skopeo inpect docker://fedora` the correct manifest
content is retrieved and a list with layers is provided. In addition,
skopeo handles the difference between schemaVersion 1 and 2.

[1] https://docs.docker.com/registry/spec/manifest-v2-2/#manifest-list-field-descriptions