safe_untar: Check for permissions to set attribs
Make sure we have permissions to restore file extended attributes.

[1] ... all processes have read access to extended security attributes,
and write access is limited to processes that have the CAP_SYS_ADMIN
capability.

[2] The file owner and processes capable of CAP_FOWNER are granted the
right to modify ACLs of a file. This is analogous to the permissions
required for accessing the file mode. (On current Linux systems, root
is the only user with the CAP_FOWNER capability.)

[1] https://linux.die.net/man/5/attr
[2] https://linux.die.net/man/1/setfacl

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>